Refactor, moved server types to their subfolders
