+++ servers/java/JavaDuplicatesHelper.py
@@ -1,0 +1,7 @@
+# TODO
+# - cache existing values
+# - warn when too many caches values
+# - get if value exists
+# - save to db when new one gets added
+# - get id from text
+# etc etc...
