+++ servers/java/JavaServer.py
@@ -7,33 +7,22 @@ import dns.resolver
 from mcstatus import JavaServer
 from mcstatus.pinger import PingResponse
 from mcstatus.status_response import JavaStatusPlayer
-from database.DbQueries import GlobalQueries, JavaDuplicateQueries, JavaQueries
+from database.DbQueries import JavaDuplicateQueries, JavaQueries
 
 from servers.Server import ServerSv
 
 from database.DbUtils import ServerType, DbUtils
+from servers.java.JavaServerFlags import JavaServerFlags
 from utils.start.startupchecks import run_startup_checks
 from utils.timer import CumulativeTimers
 from vars.DbInstances import DBINSTANCES
 from vars.DbQueues import DBQUEUES
 from vars.Errors import ERRORS, ErrorHandler
-from vars.ExceptedKeys import JAVA_EXCEPTED_KEYS
 from vars.Frontend import FRONTEND_UPDATE_THREAD
 from vars.config import Startup, Timings
 
 
-class JavaServerFlags:
-    flags_dict: dict
-    def __init__(self, table_name: str) -> None:
-        self.flags_dict = {}
 
-        exists = bool(DBINSTANCES.java_instance.cursor.execute(GlobalQueries.already_exists(table_name)).fetchone()[0])
-        if not exists:
-            return
-        for key, column_type in DBINSTANCES.java_instance.cursor.execute(GlobalQueries.get_table_info(table_name)).fetchall():
-            if key not in JAVA_EXCEPTED_KEYS.keys() or column_type == JAVA_EXCEPTED_KEYS[key]: 
-                continue
-            self.flags_dict[key] = True
 
 
 class JavaServerSv(ServerSv):
